Simplify integration plan modal lifecycle
